7/7/2017 repair login function

+++ app/Http/Controllers/StaffController.php
@@ -152,6 +152,9 @@ class StaffController extends Controller
      */
     public function postDelete($id)
     {
+        if($id == constants::ID_ADMIN){
+           return redirect('admin/staff/list')->with('note','can not delete'); 
+        }
         $staff = Staff::find($id);
         $staff->delete();
         return redirect('admin/staff/list')->with('note','delete success');
